templates/fastapi/requirements.txt
fastapi>=0.115.11
uvicorn[standard]>=0.30.0
pydantic>=2.6.0
sqlalchemy>=2.0.0
alembic>=1.13.2
psycopg[binary]>=3.2.1
httpx>=0.27.0
email-validator>=2.1.1
pytest>=8.0.0
langgraph>=0.2.72
langchain>=0.3.20
langchain-openai>=0.3.8
langchain-community>=0.3.19
python-dotenv>=1.0.1
redis>=5.0.8
azure-storage-queue>=12.11.0
